chore: refactored script management
