+++ src/app/domains/scripts/models/script-category.enum.ts
@@ -1,0 +1,9 @@
+export enum ScriptCategory {
+  Recon = 0,
+  Web = 1,
+  Crypto = 2,
+  Pwn = 3,
+  Forensics = 4,
+  Reverse = 5,
+  Misc = 6,
+}
