+++ src/app/domains/scripts/models/script.model.ts
@@ -1,0 +1,8 @@
+import { ScriptCategory } from './script-category.enum';
+
+export interface Script {
+  id: string;
+  name: string;
+  template: string;
+  category: ScriptCategory;
+}
